separate kmeans_plot and kmeans
