+++ study_gmm/kmeans.py
@@ -3,12 +3,15 @@ K-means clustering Implementation
 """
 import numpy as np
 import matplotlib.pyplot as plt
-from numpy import ndarray, random, uint8 # definition, modules
-from numpy import array, argmin, dot, ones, zeros, cov, savez # functions
 from scipy.stats import multivariate_normal
 import pickle
 from tqdm import tqdm
 import argparse
+import logging
+
+import kmeans_plot
+
+logger = logging.getLogger(__name__)
 
 class KmeansCluster():
     """Definition of clusters for K-means altorithm
@@ -104,7 +107,7 @@ class KmeansCluster():
     def train_vars_mode(self) -> str:
         return list(["outside", 'inside'])[self._train_mode]
 
-    def distortion_measure(self, x:ndarray, r:ndarray) -> float:        
+    def distortion_measure(self, x:np.ndarray, r:np.ndarray) -> float:
         """Calculate distortion measure.
 
         Args:
@@ -120,10 +123,10 @@ class KmeansCluster():
             return 0.0
         for n in range(n_sample):
             dist = [sum(v*v for v in x[n, :] - self.Mu[k, :]) for k in range(self._K)]
-            J = J + dot(r[n, :], dist)
+            J = J + np.dot(r[n, :], dist)
         return J/n_sample
 
-    def get_alignment(self, x:ndarray) -> ndarray:
+    def get_alignment(self, x: np.ndarray) -> np.ndarray:
         """
         Hard allocation of each sample to a cluster (or Gaussians)
 
@@ -137,13 +140,13 @@ class KmeansCluster():
         N = x.shape[0]
         r = np.zeros((N, self._K))
         for n in range(N):
-            r[n, argmin([ sum(v*v for v in x[n, :] - self.Mu[k, :]) for k in range(self._K)])] = 1
+            r[n, np.argmin([ sum(v*v for v in x[n, :] - self.Mu[k, :]) for k in range(self._K)])] = 1
             r[n, :] = r[n,:]/r[n,:].sum()
             #wk = [(x[n, 0] - mu[k, 0])**2 + (x[n, 1] - mu[k, 1])**2 for k in range(K)]
             #r[n, argmin(wk)] = 1
         return r
 
-    def get_soft_alignment(self, x:ndarray) -> ndarray:
+    def get_soft_alignment(self, x: np.ndarray) -> np.ndarray:
         """
         Soft alocation of each sample to Gaussians. E-step of EM algorithm to GMM.
 
@@ -162,7 +165,7 @@ class KmeansCluster():
             r[n,:] = r[n,:] / sum(r[n,:])
         return r
 
-    def get_centroid(self, x:ndarray, r:ndarray) -> ndarray:
+    def get_centroid(self, x: np.ndarray, r: np.ndarray) -> np.ndarray:
         """Calculate centroid from given alignment.
 
         Args:
@@ -173,7 +176,7 @@ class KmeansCluster():
             ndarray: updated centroid(K,D)
         """
         #  mu[k, :] = np.sum_n(r[n, k] * x[n,:]) / sum_{n,k'}(r[n, k'])
-        _mu = dot(r.transpose(), x)
+        _mu = np.dot(r.transpose(), x)
         for k in range(self._K):
             if sum(r[:,k]) < 0.01:
                 # this cluster has no sample aligned.
@@ -183,7 +186,7 @@ class KmeansCluster():
         return _mu
 
     def PushSample(self, x: np.ndarray) -> (int, float):
-        """_summary_
+        """Push one training sample to inner training variables
 
         Args:
             x (ndarray): traiing sample (D,)
@@ -194,7 +197,7 @@ class KmeansCluster():
         if self._train_mode != self.TRAIN_VAR_INSIDE:
             raise Exception('train mode is not TRAIN_VAR_INSIDE.')
         costs = [ sum([v*v for v in (x - self.Mu[k, :])]) for k in range(self._K)]
-        k_min = argmin(costs)
+        k_min = np.argmin(costs)
         self._loss += costs[k_min]
         self._X0[k_min] += 1
         self._X1[k_min,:] += x
@@ -205,6 +208,11 @@ class KmeansCluster():
         return k_min
 
     def ClearTrainigVariables(self):
+        """Reset inside statistics for training
+
+        Raises:
+            Exception: invalid triaing setting
+        """
         if self._train_mode != self.TRAIN_VAR_INSIDE:
             raise Exception('train mode is not TRAIN_VAR_INSIDE.')
         self._loss = 0.0
@@ -227,60 +235,6 @@ class KmeansCluster():
             if self._cov_mode in [KmeansCluster.COV_FULL, KmeansCluster.COV_DIAG]:
                 self.Sigma = self._X2 /self._X0[k]
         return self._loss, self._X0
-
-
-def plot_training_samples(ax, x):
-    X_range_x1 = [min(x[:,0]), max(x[:,0])]
-    X_range_x2 = [min(x[:,1]), max(x[:,1])]
-    ax.plot(x[:, 0], x[:, 1], marker='.',
-                #markerfacecolor=X_col[k],
-                markeredgecolor='k',
-                markersize=6, alpha=0.5, linestyle='none')
-    ax.set_xlim(X_range_x1)
-    ax.set_ylim(X_range_x2)
-    ax.grid("True")
-    ax.set_xlabel("$x_1$")
-    ax.set_ylabel("$x_2$")
-    ax.set_aspect("equal")
-
-
-def plot_distorion_history(cost_history: list):
-    fig, ax = plt.subplots(1, 1, figsize=(8,4))
-    ax.plot(range(0, len(cost_history)), cost_history, color='black', linestyle='-', marker='o')
-    ax.set_yscale("log")
-    ax.set_xlim([0,len(cost_history)])
-    ax.grid(True)
-    ax.set_xlabel("Iteration Step")
-    ax.set_ylabel("Loss")
-    return fig
-
-
-def show_prm(ax, x, r, mu, kmeans_param_ref:dict = {}):
-    #X_col = ['cornflowerblue', "orange", 'black', 'white', 'red']
-    K = mu.shape[0]
-    for k in range(K):
-        ax.plot(x[r[:, k] == 1, 0], x[r[:, k] == 1, 1],
-                marker='.',
-                #markerfacecolor=X_col[k],
-                markeredgecolor='k',
-                markersize=6, alpha=0.5, linestyle='none')
-    mu_ref = kmeans_param_ref.get("Mu")
-    if mu_ref is not None:
-        for k in range(K):
-            ax.plot(mu_ref[k, 0], mu_ref[k, 1], marker='*',
-                #markerfacecolor=X_col[k],
-                markersize=8, markeredgecolor='k', markeredgewidth=2)
-    for k in range(K):
-        ax.plot(mu[k, 0], mu[k, 1], marker='o',
-                #markerfacecolor=X_col[k],
-                markersize=10,
-                markeredgecolor='k', markeredgewidth=1)
-    X_range_x1 = [min(x[:,0]), max(x[:,0])]
-    X_range_x2 = [min(x[:,1]), max(x[:,1])]
-    ax.set_xlim(X_range_x1)
-    ax.set_ylim(X_range_x2)
-    ax.grid(True)
-    ax.set_aspect("equal")
 
 
 def kmeans_clustering(X:np.ndarray, mu_init:np.ndarray, max_it:int = 20, save_ckpt:bool = False):
@@ -309,70 +263,14 @@ def kmeans_clustering(X:np.ndarray, mu_init:np.ndarray, max_it:int = 20, save_ck
         loss, align_dist = kmeansparam.UpdateParameters()
         kmeansparam.ClearTrainigVariables()
         #print(loss/N, kmeansparam.Mu)
+        logger.info("iteration %d loss %f", it, loss/N)
         cost_history.append(loss)
         alignment_history.append(align_dist)
     return kmeansparam, cost_history
 
 
-def kmeans_clustering_orig(X:np.ndarray, mu_init:np.ndarray, max_it:int = 20, save_ckpt:bool = False):
-    """Run k-means clustering.
-
-    Args:
-        X (np.ndarray): vector samples (N, D)
-        mu_init (np.ndarray): initial mean vectors(K, D)
-        max_it (int, optional): Iteration steps. Defaults to 20.
-
-    Returns:
-        _type_: _description_
-    """
-    K = mu_init.shape[0]
-    Dim = X.shape[1]
-    kmeansparam = KmeansCluster(K, Dim)
-    kmeansparam.Mu = mu_init
-    R = None # alignmnet
-    cost_history = []
-    fig0, axes0 = plt.subplots(1, 4, figsize=(16, 4))
-    for it in range(0, max_it):
-        R_new = kmeansparam.get_alignment(X) # alignment to all sample to all clusters
-        #R_new = kmeansparam.get_soft_alignment(X) # not implemented yet
-        #if it > 2:
-        #    num_updated = (R_new - R) == np.zeros([N,K])
-        #    print('num_updated = ', num_updated)
-            #if num_updated == 0:
-            #    break
-        #print(R_new)
-        R = R_new
-        loss = kmeansparam.distortion_measure(X, R)
-        cost_history.append(loss)
-#        if it < (2*2):
-#            print(int(floor(it/2)), it%2)
-#            ax =axes0[it]
-            #ax =axes0[floor(it/2), it%2]
-            #show_prm(ax, X, R, mu, X_col, KmeansParam)
-#            show_prm(ax, X, R, kmeansparam.Mu)
-#            ax.set_title("iteration {0:d}".format(it + 1))
-        print(f'iteration={it} distortion={loss}')
-        kmeansparam.Mu = kmeansparam.get_centroid(X, R)
-        loss = kmeansparam.distortion_measure(X, R)
-        cost_history.append(loss)
-        if save_ckpt:
-            ckpt_file = f"kmeans_itr{it}.ckpt"
-            with open(ckpt_file, 'wb') as f:
-                pickle.dump({'model': kmeansparam,
-                    'loss': loss,
-                    'iteration': it,
-                    'alignment': R}, f)
-            print(ckpt_file)
-        # convergence check must be done
-    fig0.savefig("iteration.png")
-
-    print("centroid:", np.round(kmeansparam.Mu, 5))
-    print(kmeansparam.distortion_measure(X, kmeansparam.get_alignment(X)))
-    return kmeansparam, cost_history
-
-
 def main(args):
-    
+    print(args.input_file)
     with open(args.input_file, 'rb') as f:
         data = pickle.load(f)
         print(data.keys())
@@ -383,8 +281,8 @@ def main(args):
         
     if args.num_cluster > 0:
         Dim = X.shape[1]
-        random.seed(args.random_seed)
-        mu_init = random.randn(args.num_cluster, Dim)
+        np.random.seed(args.random_seed)
+        mu_init = np.random.randn(args.num_cluster, Dim)
 
     kmeansparam, cost_history = kmeans_clustering(X, mu_init)
     #kmeansparam, cost_history = train_light(X, mu_init)
@@ -392,8 +290,9 @@ def main(args):
     print('Sigma:', kmeansparam.Sigma)
 
     out_pngfile = "distortion.png"
-    fig = plot_distorion_history(cost_history)
+    fig = kmeans_plot.plot_distortion_history(cost_history)
     fig.savefig(out_pngfile)
+    logger.info('out: %s', out_pngfile)
 
     R = kmeansparam.get_alignment(X)
     out_file = 'out_kmeans.pickle'
@@ -403,8 +302,8 @@ def main(args):
                     'iteration': len(cost_history),
                     'alignment': R},
                     f)
-    print(out_file)
     print(sum(R==1))
+    logger.info('out: %s', out_file)
 
 
 def set_parser():
@@ -419,6 +318,11 @@ def set_parser():
     return parser
 
 if __name__ == '__main__':
+    logging.basicConfig(filename='kmeans.log',\
+        format='%(asctime)s [%(levelname)s] %(name)s %(message)s',\
+        datefmt='%m/%d/%Y %I:%M:%S', \
+        level=logging.INFO)
+
     parser = set_parser()
     args = parser.parse_args()
     main(args)
